code/requirements.txt
pygame==2.1.2
pygame-spritesheet==0.2.0